+++ packages/phorm_sqlite/lib/phorm_sqlite.dart
@@ -1,5 +1,6 @@
 export 'package:phorm/phorm.dart';
 
+export 'src/auto_migrator.dart';
 export 'src/database_adapter.dart';
 export 'src/db.dart';
 export 'src/sql_function.dart';